Change the representation of the alex_accept elements
Previously it was a [AlexAcc a user] and AlexAcc has constructors for
accept, skip, and predicated accept and skip. Then alex_scan_tkn had
to go through this list. But in fact the only times the list has more
than one element is when the list starts with predicated accept or
skip. An unconditional accept or skip always terminates the search in
alex_scan_tkn. So we can use a slightly more specialised
representation. This both saves one indirection and makes the search
function simpler.

The real motivation here is to make the search function simpler because
that will allow for greater performance improvements in the common
special case that the lexer uses no left or right contexts. In that
case the "search" actually becomes non-recursive.