Removes ignorePendingBytes for ByteString lexers.
The use of ignorePendingBytes in the ByteString
version of alexMonadScan causes an error when
the monad-bytestring or monadUserState-bytestring
wrappers are used.

This is because ignorePendingBytes is not defined,
and since it doesn't appear to be needed for
ByteString based lexers, this change removes the
use of the function from the ByteString version
of alexMonadScan.