Make build more robust on systems lacking Alex
I recently submitted a patch to Happy [1] which improves the way it
packages autogenerated .hs files in order to sidestep circular
dependency issues.

This is an adaptation of that patch to Alex, which has similar problems
in the way it handles its own circular dependency and dependency on
Happy.

In short, it is brittle to place autogenerated files into the sdist
tarball in such a way that they will extract to exactly the place where
they will be looked for during the build process.  This is mainly
because with Cabal sandboxes, the place they will be looked for during
the build process can vary widely and need not even be a subdirectory of
the extracted tarball.

This commit adds logic to Setup.lhs to support intelligent handling of
these build artifacts, both in creating and in using them.  This allows
Alex to build in a Cabal sandbox.

[1] https://github.com/simonmar/happy/pull/23