Resurrect make-sdist.sh
Because of the changes introduced in the previous commit, cabal can
produce from alex's sources an sdist tarball which will compile out of
the box even in Cabal sandboxes on systems lacking Happy or Alex or
both.

Like Happy, Alex also has a file called "make-sdist.sh" in its repo.
Unlike Happy, though, it seems to be entirely commented out, with
a comment saying that it is "not necessary anymore, Cabal does this".
However, I wasn't able to get Cabal to create an sdist tarball
containing the autogenerated .hs files from the repo in its state at
Alex version 3.1.3, so I don't understand that comment.

There seem to be some instructions for future maintainers of Alex in
comments in this file, which I have left in place in case they are still
relevant.

Since this script will presumably only be run by the maintainer, I have
taken the liberty of making it depend on reasonably new versions of
`cabal-install`.