[ fixed #141 ] regex: allow arbitary repetitions
Previously, the `r{n,m}` and related forms were restricted to single
digit numbers `n` and `m`.

We fix this by recognizing numbers of > 1 digits as NUM token and adding
rules to the regex parsing that also allows NUMs for `n` and `m`.
Previously, only CHAR was allowed, which subsumes single digits.