correct spelling of `DFS.GForrest` (#214)
> > would it break code to change to `GForest` ?
> 
> I think it wouldn't, since `alex` is an executable only and does not export an API. 
> But also, there are no strong reasons to correct the misspelling, are there? 
> (These lines mostly haven't been touched in 19 years, so we should have some respect ... :-D)

noting the computer science studies component of graph theory, the change is indicated

Skogen och vår identitet
by Erkki Lähde
Finlands natur - 1984

Prof. Lähde recounts when as a young graduate out of forestry studies, he had the occasion to visit a village in the Finnish countryside. He noticed in the area a group of trees which to his view was overdue to be cut. But enquiring upon the local villagers of that stand of trees, the reply was that there was an old woman from the village who had liked to walk among that grove. Thus the trees were saved out of piety.

>  (These lines mostly haven't been touched in 19 years, so we should have some respect ... :-D)

piety isn't indicated here