Deduplicate docs, prepare to get rid of website
The idea is that the docs should the source of truth, and be able to
replace the website (#215).

Information from the read-me is also moved to the docs.

The `CONTRIBUTING.rst` symlink is intended to do a thing on GitHub (see
https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors),
but I am not sure how to confirm it will short of merging this PR.