Fix #228. Remove unused function Util.brack.
The function Util.brack is not used by any part of alex. A quick search
through the tree with grep shows this.

Since alex doesn't expose an API and the function is not used
internally, it might as well be removed from the module completely.

This commit removes the function brack from the module Util.