Fix #230. Remove unused functions related to insert sort.
The function Sort.isort is not used by any part of alex. A quick search
through the tree with grep shows this. The function Sort.insrt is only
ever used by Sort.isort.

Since alex doesn't expose an API and the function is not used
internally, both can be removed from the module completely.

This commit removes the functions isort and insrt from the module Sort.
The comment at the top of the module is adjusted to remove mentions of
insert sort or isort.